adding additional changes made to Porducts page

+++ models/user.js
@@ -63,7 +63,7 @@ module.exports = function (connection, Sequelize) {
       },
     },
     isAdmin: {
-      type: Sequelize.TINYINT,
+      type: Sequelize.INTEGER,
     },
   });
